Refactor: Add user package in dto
