+++ src/main/java/com/alpaco/dev/dto/user/UserSignInRequestDto.java
@@ -1,4 +1,4 @@
-package com.alpaco.dev.dto;
+package com.alpaco.dev.dto.user;
 
 import lombok.Data;
