+++ src/main/java/com/alpaco/dev/mapper/UserMapper.java
@@ -1,6 +1,6 @@
 package com.alpaco.dev.mapper;
 
-import com.alpaco.dev.dto.UserSignUpRequestDto;
+import com.alpaco.dev.dto.user.UserSignUpRequestDto;
 import com.alpaco.dev.entity.user.User;
 import org.apache.ibatis.annotations.Mapper;
